WIP: Evacuator (OpenJ9 integration -- final)
Squashing all commits following original pull request candidate commit
(2018-04-03). Changes included in this commit are documented in #2664
and #2361.

As of this commit, 64-bit builds are working and performing fine, 
consistently passing all git tests in internal x86_64, ppc and zos 
64-bit builds for linux. Recent changes to scavenger (outside of 
evacuator, possibly introduction of scavenger delegate or changes 
relating to heap resizing) have broken 32-bit builds. This bug is 
present in evacuator builds regardless of whether evacuator or scavenger
is selected as algorithm for gencon generational collector (scavenger is
default as of this commit, evacuator is selected using
`-XXgc:recursiveScanOrdering`.

The 32-bit bug can be reproduced by running an evacuator-enabled build
on specJVM2008/xml.validation with default heap size. There will be a 
segfault on NULL pointer dereference after a few GC cycles after 
concuuent marking is kicked off. When this is run with `-Xcheck:gc`,
gccheck reports no errors up to point of failure, which occurs after the
stop-the-world phase for most recent nursery collection completes. 

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>